Updated git submodules
Project: mediawiki/extensions/CollaborationKit master 44f8fbdc1433ee043052a3ca6b9390da327753cd

Purges cache upon adding a new feature to a hub.

This is so that there's no "feature does not exist" text when the hub is loaded for the first time after the new feature is created.

For hub *creation*, it just creates the subpages first to short-circuit this problem altogether.

Bug: T149022
Change-Id: I12bd045104cd61e1cad4b9725d58736a80fdd3f1