Updated git submodules
Project: mediawiki/extensions/UploadWizard master 5b8016359f069f39b9fa2a867f2a9c6b2eb4eb49

mw.UploadWizardLicenseInput: Correct unguarded for...in

A for...in loop will also interate over the properties of the
prototype of the object. If some other script adds custom properties
to Array.prototype or Object.prototype, we'll find them here,
`template[ '*' ]` will most likely be undefined and result in an
exception from mw.Title.

Instead, use a regular for loop. (Wrapping the body of the loop in a
hasOwnProperty() check would also work, but since we're iterating over
an array, this is nicer.)

Bug: T151220
Change-Id: I273b2921f7171754fe9ffac6a84cae7a3d10c05a