Updated git submodules
Project: mediawiki/extensions/Wikibase master ea969334fcdfb3a07c83285b5a4c9138440bfaca

Remove termWeight field from TermIndexEntry

This patch does *not* drop the term_weight field from the terms table.
But the field is not exposed any more via the TermIndexEntry class.
This makes the field a private implementation detail of TermSqlIndex.
The weight is still calculated, stored and used for ordering, but not
exposed any more. It was not used anyway outside of TermSqlIndex.

Change-Id: I25cf98c4dc72c52aae5bd5545066a2f577b4fb2c