Updated git submodules
Project: mediawiki/extensions/ORES master 92951715dfc94c0c3f9e839c4d2df947448271fc

'goodfaith' filter on Special:RC / Special:Watchlist

This filter is a not a typical show/hide toggle.
It is url-driven only for the moment and will be used
by the new RC filters (ERI project).

It accepts one or many of the following values:
  good, maybebad, bad
Values are separated by a comma.

Each value is associated with a range of probabilities
that are compared against the goodfaith test. They
can be configured using $wgOresGoodfaithLevels.

Bug: T149853
Depends-On: If47842f3d91c5999a7c5bf25666b967e9b30a6d7
Change-Id: I28c593f34145d566f803d96ff8220fd685c2f695