Updated git submodules
Project: mediawiki/extensions/RevisionSlider master bf55a214e5e26c232dd83114195a55a7d9869c6a

Remove doubled padding

The inline padding on the widget element seems to be the original reason
for the flickering. To reflect the CSS it needed to be moved to the
button element.

I overlooked this when fixing the lazy load patch. 
https://gerrit.wikimedia.org/r/#/c/323193/16..17/RevisionSlider.hooks.php

This patch fixes it by removing the now doubled padding on the 
widget element.

Change-Id: Ie904f619e8c84474653d2ba40d9f9e62e74fa9db