Remove Widgets extension as submodule from meta repo
The extension is currently broken in git, and updating this repo
leaves one with a dirty clone that cannot be cleaned up. The error
is as follows:

error: Server does not allow request for unadvertised object 35480f10e7ce9b0fdaf23d3799d7b79463919b1e
Fetched in submodule path 'smarty', but it did not contain 35480f10e7ce9b0fdaf23d3799d7b79463919b1e. Direct fetching of that commit failed.

It has to do with the submodule in question that it makes use of.
If someone can fix the repo, this change can be reverted.

Change-Id: I1ad244f34b3eac0c6094ba77efa8c703ef7ae96c