sync-with-gerrit: expand ls-projects options
The `ls-projects` command has a few options and I had to lookup what
`-d` was for: it is to include the projects descriptions.

Replace the short description by their long alternatives which have the
benefit of being self describing.
Split the list to have one option/arg per line so that a later addition
shows up line being added/removed and thus ease reviewing.

Bug: T378301
Change-Id: Id4e68d935df42e5fdc6144a11b91754e89dc2f86